requirements.txt
flask>=1.1.2
gunicorn>=19.7.0
gevent>=1.4
sqlalchemy>=1.3.20
pyyaml>=5.3.1
#psycopg2>=2.8.6
flask_sqlalchemy>=2.4.4
Flask-Migrate>=2.5.3
